Fully replace VolumeCloudRenderer.RenderFrame
This change replaces the existing RenderFrame method with a custom one
meant to fix issues with stereoscopic rendering. The main changes to the
original algorithm include:
* Get projection matrices approrpriate for currently rendered eye
* Keep track of separate state-maintaining variables per eye, such as
  buffers and indexes.
* Does not support spherical map rendering, which AFAICT is not used by
  Subnautica anyways.

I tested this on a Valve Index with SteamVR and an Oculus Quest 2 using
Air Link with both OpenVR and OculusVR modes as well as streaming via
Virtual Desktop with OpenVR mode.