Fix camera positioning (#13)
Without this patch, the camera is positioned just behind the player's
neck normally. When the player would pull out the PDA, the camera shifts
forward to the centered position on the player's neck. This patch
removes that behavior so that the camera is always positioned on the
player neck even when the PDA is not active.

This change also bypasses some code that sets a clipping value that was
causing the player body to be clipped nearby, so now looking down at the
body won't cause clipping visuals and instead you'll see the player body
unless you move your head actually inside of the body.